projects-demo/whisper-translates/requirements.txt
streamlit
st_audiorec
openai-whisper
pydub